requirements.txt
django==4.0
google-api-python-client==2.33
google-auth
google-auth-oauthlib==0.4.6
google-auth-httplib2
redis==4.0.2
